requirements.txt
pygame
pyopengl
numpy
pyglm
pywavefront
